// src/generated/wotlk/common.ts
export enum Stat {
    /**
     * @generated from protobuf enum value: StatStrength = 0;
     */
    StatStrength = 0,
    /**
     * @generated from protobuf enum value: StatAgility = 1;
     */
    StatAgility = 1,
    /**
     * @generated from protobuf enum value: StatStamina = 2;
     */
    StatStamina = 2,
    /**
     * @generated from protobuf enum value: StatIntellect = 3;
     */
    StatIntellect = 3,
    /**
     * @generated from protobuf enum value: StatSpirit = 4;
     */
    StatSpirit = 4,
    /**
     * @generated from protobuf enum value: StatSpellPower = 5;
     */
    StatSpellPower = 5,
    /**
     * @generated from protobuf enum value: StatMP5 = 6;
     */
    StatMP5 = 6,
    /**
     * @generated from protobuf enum value: StatSpellHit = 7;
     */
    StatSpellHit = 7,
    /**
     * @generated from protobuf enum value: StatSpellCrit = 8;
     */
    StatSpellCrit = 8,
    /**
     * @generated from protobuf enum value: StatSpellHaste = 9;
     */
    StatSpellHaste = 9,
    /**
     * @generated from protobuf enum value: StatSpellPenetration = 10;
     */
    StatSpellPenetration = 10,
    /**
     * @generated from protobuf enum value: StatAttackPower = 11;
     */
    StatAttackPower = 11,
    /**
     * @generated from protobuf enum value: StatMeleeHit = 12;
     */
    StatMeleeHit = 12,
    /**
     * @generated from protobuf enum value: StatMeleeCrit = 13;
     */
    StatMeleeCrit = 13,
    /**
     * @generated from protobuf enum value: StatMeleeHaste = 14;
     */
    StatMeleeHaste = 14,
    /**
     * @generated from protobuf enum value: StatArmorPenetration = 15;
     */
    StatArmorPenetration = 15,
    /**
     * @generated from protobuf enum value: StatExpertise = 16;
     */
    StatExpertise = 16,
    /**
     * @generated from protobuf enum value: StatMana = 17;
     */
    StatMana = 17,
    /**
     * @generated from protobuf enum value: StatEnergy = 18;
     */
    StatEnergy = 18,
    /**
     * @generated from protobuf enum value: StatRage = 19;
     */
    StatRage = 19,
    /**
     * @generated from protobuf enum value: StatArmor = 20;
     */
    StatArmor = 20,
    /**
     * @generated from protobuf enum value: StatRangedAttackPower = 21;
     */
    StatRangedAttackPower = 21,
    /**
     * @generated from protobuf enum value: StatDefense = 22;
     */
    StatDefense = 22,
    /**
     * @generated from protobuf enum value: StatBlock = 23;
     */
    StatBlock = 23,
    /**
     * @generated from protobuf enum value: StatBlockValue = 24;
     */
    StatBlockValue = 24,
    /**
     * @generated from protobuf enum value: StatDodge = 25;
     */
    StatDodge = 25,
    /**
     * @generated from protobuf enum value: StatParry = 26;
     */
    StatParry = 26,
    /**
     * @generated from protobuf enum value: StatResilience = 27;
     */
    StatResilience = 27,
    /**
     * @generated from protobuf enum value: StatHealth = 28;
     */
    StatHealth = 28,
    /**
     * @generated from protobuf enum value: StatArcaneResistance = 29;
     */
    StatArcaneResistance = 29,
    /**
     * @generated from protobuf enum value: StatFireResistance = 30;
     */
    StatFireResistance = 30,
    /**
     * @generated from protobuf enum value: StatFrostResistance = 31;
     */
    StatFrostResistance = 31,
    /**
     * @generated from protobuf enum value: StatNatureResistance = 32;
     */
    StatNatureResistance = 32,
    /**
     * @generated from protobuf enum value: StatShadowResistance = 33;
     */
    StatShadowResistance = 33,
    /**
     * @generated from protobuf enum value: StatBonusArmor = 34;
     */
    StatBonusArmor = 34,
    /**
     * Placeholder: not used yet.
     *
     * @generated from protobuf enum value: StatRunicPower = 35;
     */
    StatRunicPower = 35,
    /**
     * Placeholder: not used yet.
     *
     * @generated from protobuf enum value: StatBloodRune = 36;
     */
    StatBloodRune = 36,
    /**
     * Placeholder: not used yet.
     *
     * @generated from protobuf enum value: StatFrostRune = 37;
     */
    StatFrostRune = 37,
    /**
     * Placeholder: not used yet.
     *
     * @generated from protobuf enum value: StatUnholyRune = 38;
     */
    StatUnholyRune = 38,
    /**
     * Placeholder: not used yet.
     *
     * @generated from protobuf enum value: StatDeathRune = 39;
     */
    StatDeathRune = 39
}
// ...
export enum ItemType {
    /**
     * @generated from protobuf enum value: ItemTypeUnknown = 0;
     */
    ItemTypeUnknown = 0,
    /**
     * @generated from protobuf enum value: ItemTypeHead = 1;
     */
    ItemTypeHead = 1,
    /**
     * @generated from protobuf enum value: ItemTypeNeck = 2;
     */
    ItemTypeNeck = 2,
    /**
     * @generated from protobuf enum value: ItemTypeShoulder = 3;
     */
    ItemTypeShoulder = 3,
    /**
     * @generated from protobuf enum value: ItemTypeBack = 4;
     */
    ItemTypeBack = 4,
    /**
     * @generated from protobuf enum value: ItemTypeChest = 5;
     */
    ItemTypeChest = 5,
    /**
     * @generated from protobuf enum value: ItemTypeWrist = 6;
     */
    ItemTypeWrist = 6,
    /**
     * @generated from protobuf enum value: ItemTypeHands = 7;
     */
    ItemTypeHands = 7,
    /**
     * @generated from protobuf enum value: ItemTypeWaist = 8;
     */
    ItemTypeWaist = 8,
    /**
     * @generated from protobuf enum value: ItemTypeLegs = 9;
     */
    ItemTypeLegs = 9,
    /**
     * @generated from protobuf enum value: ItemTypeFeet = 10;
     */
    ItemTypeFeet = 10,
    /**
     * @generated from protobuf enum value: ItemTypeFinger = 11;
     */
    ItemTypeFinger = 11,
    /**
     * @generated from protobuf enum value: ItemTypeTrinket = 12;
     */
    ItemTypeTrinket = 12,
    /**
     * @generated from protobuf enum value: ItemTypeWeapon = 13;
     */
    ItemTypeWeapon = 13,
    /**
     * @generated from protobuf enum value: ItemTypeRanged = 14;
     */
    ItemTypeRanged = 14
}
/**
 * @generated from protobuf enum proto.ArmorType
 */
export enum ArmorType {
    /**
     * @generated from protobuf enum value: ArmorTypeUnknown = 0;
     */
    ArmorTypeUnknown = 0,
    /**
     * @generated from protobuf enum value: ArmorTypeCloth = 1;
     */
    ArmorTypeCloth = 1,
    /**
     * @generated from protobuf enum value: ArmorTypeLeather = 2;
     */
    ArmorTypeLeather = 2,
    /**
     * @generated from protobuf enum value: ArmorTypeMail = 3;
     */
    ArmorTypeMail = 3,
    /**
     * @generated from protobuf enum value: ArmorTypePlate = 4;
     */
    ArmorTypePlate = 4
}
/**
 * @generated from protobuf enum proto.WeaponType
 */
export enum WeaponType {
    /**
     * @generated from protobuf enum value: WeaponTypeUnknown = 0;
     */
    WeaponTypeUnknown = 0,
    /**
     * @generated from protobuf enum value: WeaponTypeAxe = 1;
     */
    WeaponTypeAxe = 1,
    /**
     * @generated from protobuf enum value: WeaponTypeDagger = 2;
     */
    WeaponTypeDagger = 2,
    /**
     * @generated from protobuf enum value: WeaponTypeFist = 3;
     */
    WeaponTypeFist = 3,
    /**
     * @generated from protobuf enum value: WeaponTypeMace = 4;
     */
    WeaponTypeMace = 4,
    /**
     * @generated from protobuf enum value: WeaponTypeOffHand = 5;
     */
    WeaponTypeOffHand = 5,
    /**
     * @generated from protobuf enum value: WeaponTypePolearm = 6;
     */
    WeaponTypePolearm = 6,
    /**
     * @generated from protobuf enum value: WeaponTypeShield = 7;
     */
    WeaponTypeShield = 7,
    /**
     * @generated from protobuf enum value: WeaponTypeStaff = 8;
     */
    WeaponTypeStaff = 8,
    /**
     * @generated from protobuf enum value: WeaponTypeSword = 9;
     */
    WeaponTypeSword = 9
}
/**
 * @generated from protobuf enum proto.HandType
 */
export enum HandType {
    /**
     * @generated from protobuf enum value: HandTypeUnknown = 0;
     */
    HandTypeUnknown = 0,
    /**
     * @generated from protobuf enum value: HandTypeMainHand = 1;
     */
    HandTypeMainHand = 1,
    /**
     * @generated from protobuf enum value: HandTypeOneHand = 2;
     */
    HandTypeOneHand = 2,
    /**
     * @generated from protobuf enum value: HandTypeOffHand = 3;
     */
    HandTypeOffHand = 3,
    /**
     * @generated from protobuf enum value: HandTypeTwoHand = 4;
     */
    HandTypeTwoHand = 4
}
/**
 * @generated from protobuf enum proto.RangedWeaponType
 */
export enum RangedWeaponType {
    /**
     * @generated from protobuf enum value: RangedWeaponTypeUnknown = 0;
     */
    RangedWeaponTypeUnknown = 0,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeBow = 1;
     */
    RangedWeaponTypeBow = 1,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeCrossbow = 2;
     */
    RangedWeaponTypeCrossbow = 2,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeGun = 3;
     */
    RangedWeaponTypeGun = 3,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeIdol = 4;
     */
    RangedWeaponTypeIdol = 4,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeLibram = 5;
     */
    RangedWeaponTypeLibram = 5,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeThrown = 6;
     */
    RangedWeaponTypeThrown = 6,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeTotem = 7;
     */
    RangedWeaponTypeTotem = 7,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeWand = 8;
     */
    RangedWeaponTypeWand = 8,
    /**
     * @generated from protobuf enum value: RangedWeaponTypeSigil = 9;
     */
    RangedWeaponTypeSigil = 9
}